Needs some more work but we're almost there!
reverted to before max and sawyer were here and it fixed most of our issues. We reworked some of autonomous too and the two in speaker from position 2 works consistently. We now have the ground work for the sides. The isOnRed chooser needs to be testeed again though. WE'RE MAKING LOTS OF PROGRESS!!!